[fix] duplicated nickname error on SingUp handled
